+++ AnimatorGL/Settings.h
@@ -9,14 +9,14 @@ static const int SAVE_PNG_EVERY = 500;
 static const int DEFAULT_NUMBER_OF_THREADS = 4;
 
 static const int DEFAULT_PARTICLE_COUNT = 10;
-static const double DEFAULT_TOTAL_TIME_STEPS = 10.0;
-static const double TIME_STEP = 0.01;
-static const double MIN_DISTANCE = 10.0;
+static const float DEFAULT_TOTAL_TIME_STEPS = 10.0f;
+static const float TIME_STEP = 0.01;
+static const float MIN_DISTANCE = 10.0f;
 
-static const double GRAVITATIONAL_CONSTANT = 6.673e-11;
+static const float GRAVITATIONAL_CONSTANT = 6.673e-11;
 
-static const double MAX_RANDOM = 1.0;
-static const double MAX_MASS = 1.0;
+static const float MAX_RANDOM = 1.0f;
+static const float MAX_MASS = 1.0f;
 
 static const int UNIVERSE_SIZE_X = 800;
 static const int UNIVERSE_SIZE_Y = 800;
